backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis & Celery
redis[hiredis]==5.0.1
celery[redis]==5.3.6

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.3.0

# Security
bleach==6.1.0 

# Kalshi API
httpx==0.27.0
cryptography==42.0.0

# Payments
stripe==7.11.0

# Notifications
python-telegram-bot==22.5

# Utilities
python-dotenv==1.0.0
loguru==0.7.2